inference/requirements.txt
fastapi
uvicorn[standard]
torch
transformers
Pillow
